Hacked to use openssl instead of gnutls.
This is only a prototype to check what must be done to have a functional
version. I still have to find a way to support openssl without breaking
applications that expect gnutls. Indeed, the microhttpd API exposes gnutls
types or objects. We want to keep existing applications functional. So
we may have to add an extra API more suited to openssl use (SSL-engine
agnostic if possible) and let the application dynamically choose between
available SSL engines.